requirements.txt
# Orion — Python Dependencies
# Part of Orion — Persistent AI Companion System.

# LLM Frameworks
langchain>=0.2.0
langgraph>=0.1.0
langchain-community>=0.2.0

# LLM Providers
openai>=1.30.0
anthropic>=0.25.0
google-generativeai>=0.5.0
ollama>=0.2.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Vector Database
pinecone-client>=3.0.0
supabase>=2.0.0

# Background Processing
celery>=5.3.0
redis>=5.0.0

# Browser Automation
playwright>=1.40.0
browser-use>=0.1.0

# Desktop Automation
pyautogui>=0.9.54
opencv-python>=4.9.0

# Voice
openai-whisper>=20231117

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# File Watching
watchdog>=4.0.0

# Web Framework
fastapi>=0.110.0
uvicorn>=0.27.0

# HTTP
httpx>=0.27.0
requests>=2.31.0
